requirements.txt
discord.py==2.3.2
aiohttp==3.9.5
websockets==12.0
python-dotenv==1.0.1
# Optional: for RugCheck JWT auth with Solana private key
# solders==0.21.0
# base58==2.1.1
